move to resource model
