+++ front/lib/resources/webhook_source_resource.ts
@@ -15,12 +15,20 @@ import type { ReadonlyAttributesType } from "@app/lib/resources/storage/types";
 import { getResourceIdFromSId, makeSId } from "@app/lib/resources/string_ids";
 import { TriggerResource } from "@app/lib/resources/trigger_resource";
 import type { ResourceFindOptions } from "@app/lib/resources/types";
+import { GitHubWebhookService } from "@app/lib/triggers/services/github_webhook_service";
+import type { RemoteWebhookService } from "@app/lib/triggers/services/remote_webhook_service";
 import { DEFAULT_WEBHOOK_ICON } from "@app/lib/webhookSource";
+import logger from "@app/logger/logger";
 import type { ModelId, Result } from "@app/types";
 import { Err, normalizeError, Ok, redactString } from "@app/types";
 import type { WebhookSourceType } from "@app/types/triggers/webhooks";
 
 const SECRET_REDACTION_COOLDOWN_IN_MINUTES = 10;
+
+// Service registry: map webhook source kind to its service implementation
+const WEBHOOK_SERVICES: Record<string, RemoteWebhookService> = {
+  github: new GitHubWebhookService(),
+};
 
 // Attributes are marked as read-only to reflect the stateless nature of our Resource.
 // This design will be moved up to BaseResource once we transition away from Sequelize.
@@ -172,6 +180,30 @@ export class WebhookSourceResource extends BaseResource<WebhookSourceModel> {
     );
 
     const owner = auth.getNonNullableWorkspace();
+
+    const service = WEBHOOK_SERVICES[this.kind];
+    if (service && this.remoteMetadata && this.oauthConnectionId) {
+      try {
+        const result = await service.deleteWebhooks({
+          auth,
+          connectionId: this.oauthConnectionId,
+          remoteMetadata: this.remoteMetadata,
+        });
+
+        if (result.isErr()) {
+          logger.error(
+            `Failed to delete remote webhook on ${this.kind}`,
+            result.error.message
+          );
+        }
+      } catch (error) {
+        logger.error(
+          `Failed to delete remote webhook on ${this.kind}`,
+          error instanceof Error ? error.message : error
+        );
+        // Continue with local deletion even if remote deletion fails
+      }
+    }
 
     try {
       // Find all webhook sources views for this webhook source
