+++ front/pages/api/w/[wId]/webhook_sources/[webhookSourceId]/index.ts
@@ -3,16 +3,8 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { WebhookSourceResource } from "@app/lib/resources/webhook_source_resource";
-import { GitHubWebhookService } from "@app/lib/triggers/services/github_webhook_service";
-import type { RemoteWebhookService } from "@app/lib/triggers/services/remote_webhook_service";
-import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
-
-// Service registry: map webhook source kind to its service implementation
-const WEBHOOK_SERVICES: Record<string, RemoteWebhookService> = {
-  github: new GitHubWebhookService(),
-};
 
 export type DeleteWebhookSourceResponseBody = {
   success: true;
@@ -111,35 +103,6 @@ async function handler(
                 "The webhook source you're trying to delete was not found.",
             },
           });
-        }
-
-        // If this webhook has remote metadata and a service exists for this kind, try to delete it from the provider
-        const service = WEBHOOK_SERVICES[webhookSourceResource.kind];
-        if (
-          service &&
-          webhookSourceResource.remoteMetadata &&
-          webhookSourceResource.oauthConnectionId
-        ) {
-          try {
-            const result = await service.deleteWebhooks({
-              auth,
-              connectionId: webhookSourceResource.oauthConnectionId,
-              remoteMetadata: webhookSourceResource.remoteMetadata,
-            });
-
-            if (result.isErr()) {
-              logger.error(
-                `Failed to delete remote webhook on ${webhookSourceResource.kind}`,
-                result.error.message
-              );
-            }
-          } catch (error) {
-            logger.error(
-              `Failed to delete remote webhook on ${webhookSourceResource.kind}`,
-              error instanceof Error ? error.message : error
-            );
-            // Continue with local deletion even if remote deletion fails
-          }
         }
 
         const deleteResult = await webhookSourceResource.delete(auth);
